Bypass casting in ToRadians and ToDegrees

+++ MonoGame.Framework/MathHelper.cs
@@ -272,7 +272,7 @@ namespace Microsoft.Xna.Framework
         /// </remarks>
         public static float ToDegrees(float radians)
         { 
-            return (float)(radians * 57.295779513082320876798154814105);
+            return (radians * 57.295779513082320876798154814105f);
         }
 
         /// <summary>
@@ -300,7 +300,7 @@ namespace Microsoft.Xna.Framework
         /// </remarks>
         public static float ToRadians(float degrees)
         { 
-            return (float)(degrees * 0.017453292519943295769236907684886);
+            return (degrees * 0.017453292519943295769236907684886f);
         }
 
         /// <summary>
